requirements.txt
git+https://github.com/CityOfPhiladelphia/ais.git#egg=ais
git+https://github.com/CityOfPhiladelphia/datum.git#egg=datum
git+https://github.com/CityOfPhiladelphia/geopetl.git#egg=geopetl
git+https://github.com/CityOfPhiladelphia/passyunk
git+ssh://git@private-git/CityOfPhiladelphia/passyunk_automation.git
alembic==1.8.1
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
banal==1.0.6
blinker==1.5
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
config==0.5.1
cx-Oracle==8.3.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.2.1
exceptiongroup==1.0.4
executing==1.2.0
Flask==2.2.2
Flask-CacheControl==0.3.0
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
fuzzywuzzy==0.18.0
GeoAlchemy2==0.12.5
geopetl==0.0.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
guppy3==3.1.2
idna==3.4
importlib-metadata==6.1.0
iniconfig==1.1.1
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Levenshtein==0.20.8
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
normality==2.4.0
numpy==1.24.2
packaging==21.3
pandas==1.5.3
parso==0.8.3
pathlib2==2.3.7.post1
petl==1.7.12
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycogreen==1.0.2
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyproj==3.4.0
pytest==7.2.0
python-dateutil==2.8.1
python-dotenv==0.21.0
python-editor==1.0.4
python-Levenshtein==0.20.8
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.2
raven==6.10.0
requests==2.28.1
Shapely==1.8.5.post1
six==1.10.0
SQLAlchemy==1.4.44
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.5.2